app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
pydantic==2.9.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.2.0
python-dotenv==1.0.1
